projects/content-system/demos/fnol-voice-agent/requirements.txt
livekit-agents==1.3.10
livekit-plugins-openai==1.3.10
livekit-plugins-cartesia==1.3.10
livekit-plugins-deepgram==1.3.10
livekit-plugins-silero==1.3.10
livekit-plugins-turn-detector==1.3.10
python-dotenv>=1.0.0
